+++ NEWS.md
@@ -1,10 +1,12 @@
 # odbc (development version)
 
-* SQL Server: Support for `DATETIMEOFFSET` data type (#918). 
-
 * Fix hang when parsing exceptionally long database errors (#916).
 
 * DB2: Fix error when writing to temp tables (#910).
+
+* DB2: Add support for `DB2_XML` (#913).
+
+* SQL Server: Support for `DATETIMEOFFSET` data type (#918).
 
 * SQL Server: Fix issues with argument propagation in bespoke
   S4 methods (#906).